+++ cpp/glfw_imgui/src/vectormath.h
@@ -11,31 +11,31 @@
 
 struct Vector
 {
-	float x, y, z;
+	double x, y, z;
 
 	Vector();
 	Vector(const Vector& v);
-	Vector(float x, float y, float z);
-	Vector(float f);
+	Vector(double x, double y, double z);
+	Vector(double f);
 
 	virtual ~Vector();
 
-	float length2();
-	float length();
+	double length2();
+	double length();
 
 	Vector normal();
-	float normal_len();
+	double normal_len();
 	
 
 	Vector& operator =(const Vector& v);
 	Vector& operator +=(const Vector& v);
 	Vector& operator -=(const Vector& v);
-	Vector& operator *=(float f);
-	Vector& operator /=(float f);
+	Vector& operator *=(double f);
+	Vector& operator /=(double f);
 	Vector operator -() const;
 };
 
-float dot(Vector v1, Vector v2);
+double dot(Vector v1, Vector v2);
 Vector cross(Vector v1, Vector v2);
 
 inline Vector operator +(const Vector& v1, const Vector& v2)
@@ -59,14 +59,14 @@ inline Vector operator *(const Vector& v1, const Vector& v2)
 		v1.z * v2.z);
 }
 
-inline Vector operator *(const Vector& v, float f)
+inline Vector operator *(const Vector& v, double f)
 {
 	return Vector(v.x * f,
 		v.y * f,
 		v.z * f);
 }
 
-inline Vector operator *(float f, const Vector& v)
+inline Vector operator *(double f, const Vector& v)
 {
 	return Vector(f * v.x,
 		f * v.y,
@@ -80,14 +80,14 @@ inline Vector operator /(const Vector& v1, const Vector& v2)
 		v1.z / v2.z);
 }
 
-inline Vector operator /(const Vector& v, float f)
+inline Vector operator /(const Vector& v, double f)
 {
 	return Vector(v.x / f,
 		v.y / f,
 		v.z / f);
 }
 
-inline Vector operator /(float f, const Vector& v)
+inline Vector operator /(double f, const Vector& v)
 {
 	return Vector(f / v.x, 
 		f / v.y,
@@ -98,12 +98,12 @@ typedef Vector Point;
 
 struct Vector2
 {
-	float u, v;
+	double u, v;
 
 	Vector2();
 	Vector2(const Vector2& v);
-	Vector2(float u, float v);
-	Vector2(float f);
+	Vector2(double u, double v);
+	Vector2(double f);
 
 	virtual ~Vector2();
